Time: 3 ms (38.74%) | Memory: 17.3 MB (15.07%) - LeetSync

+++ 150-evaluate-reverse-polish-notation/evaluate-reverse-polish-notation.cpp
@@ -20,6 +20,6 @@ public:
             st.push(stoll(it)); 
         }
         }
-        return static_cast<int>(st.top());
+        return st.top();
     }
 };
